create more service logic in transaction

+++ entity/transaction.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Transaction struct {
 	ID         int64          `gorm:"primaryKey" json:"id"`
 	Invoice    string         `json:"invoice" validate:"required"`
-	Qty        int64          `json:"qty"`
+	Qty        float64        `json:"qty"`
 	StatusId   int            `json:"status_id"`
 	Status     Status         `json:"status"`
 	CustomerId int            `json:"customer_id"`
